requirements.txt
deepface
fastapi[standard]
numpy
tf-keras
pgvector
sqlmodel
tensorflow
python-dotenv